requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
chromadb
pypdf
gradio
python-dotenv
